conductor: "Add controller. WARN: linting failed.. WARN: extra files found, committing them. "
