+++ pkg/controller/direct/bigtable/client.go
@@ -37,7 +37,7 @@ func newGCPClient(ctx context.Context, config *config.ControllerConfig) (*gcpCli
 }
 
 func (m *gcpClient) newBigtableInstanceAdminClient(ctx context.Context) (*api.InstanceAdminClient, error) {
-	opts, err := m.config.RESTClientOptions()
+	opts, err := m.config.GRPCClientOptions()
 	if err != nil {
 		return nil, err
 	}
